#10 [WordCloud] implemented Wordcloud action and opens dialog
implemented the action for generating a wordcloud. Added a button for the action to the toolbar using a could icon for visuals. Made sure the user has to select entries to be able to generate a wordcloud. commented what needs to be done for the panel in the dialog and the generation of the wordcloud.